requirements.txt
Flask==3.0.3
openai>=1.40.0
pypdf==4.3.1
python-dotenv==1.0.1
gunicorn==22.0.0
